requirements.txt
pyinstaller==4.3
requests==2.25.1
tqdm==4.61.1
colorama==0.4.4
pycryptodome==3.10.1